Update the Wiki with PS5 changes
DH made a new WSL script today for ease of installation to help Windows people with getting started. So I incorporated that script into the install guide and decided to run through every file and cleanup what I could see was notably functional but could be improved wording, URLs, and more.
PlayStation 5 support was added to RPCSX and there is a lot more working than just 2 titles at this time. Many games are seeing title screens and commands are overdue on simplification for booting content.